requirements.txt
torch>=2.1.0
transformers>=4.40.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2
ollama>=0.2.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
rich>=13.7.0
ragas>=0.1.9
datasets>=2.19.0
pytest>=8.0.0